+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- **Pagination** `Button` to `ButtonWithIcon`
+
 ## [8.30.1] - 2019-04-10
 
 ### Fixed